feat(java-concurrent): Introduce dedicated instrumentation module
WIP
